requirements.txt
streamlit
python-dotenv
Pillow
google-generativeai
pandas
tenacity
tqdm